+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `meta` as export format
+
 ### Fixed
 
 - `xlsx` importer did not give the correct node name to segmentation token. Due